feat. 수동으로 입력 받은 것을 저장하고, 자동으로 구매할 개수를 계산한다
