+++ README.md
@@ -165,4 +165,4 @@
 
 - [x] 수동으로 구매할 로또 수를 입력 받는다 - InputView
 - [x] 수동으로 구매할 번호를 입력 받는다 - InputView
-- [ ] 수동으로 입력 받은 것을 저장하고, 자동으로 구매할 개수를 계산한다 - Tickets
+- [x] 수동으로 입력 받은 것을 저장하고, 자동으로 구매할 개수를 계산한다 - Tickets
