+++ src/main/kotlin/lotto/controller/Main.kt
@@ -12,12 +12,12 @@ fun main() {
     LotteryShop.validateManualNum(manualNum, lottoCount)
     val manualLottos = InputView.getManualLotto(manualNum)
     val tickets = LotteryShop.getTickets(lottoCount, manualLottos)
-    ResultView.printLottoNums(tickets)
+    ResultView.printLottos(manualNum, tickets)
 
     val lastLottoNums = InputView.getWinningLottoNums()
     val bonusNum = InputView.getBonusNum()
     val winningTicket = WinningTicket(lastLottoNums, bonusNum)
     val score = winningTicket.score(tickets)
-    val rate = winningTicket.calculateRateOfReturn(money, score)
+    val rate = LotteryShop.calculateRateOfReturn(money, score)
     ResultView.printResult(score, rate)
 }
