+++ src/main/kotlin/lotto/domain/Tickets.kt
@@ -1,8 +1,13 @@
 package lotto.domain
 
-class Tickets(lottoCount: Int, private val numCreator: LottoNumGenerator) {
+class Tickets(lottoCount: Int, manualLotto: List<Lotto>, private val numCreator: LottoNumGenerator) {
 
-    val tickets = List(lottoCount) { Lotto(getLottoNumber()) }
+    val tickets: List<Lotto>
+
+    init {
+        val autoLottos = List(lottoCount - manualLotto.size) { Lotto(getLottoNumber()) }
+        tickets = manualLotto + autoLottos
+    }
 
     private fun getLottoNumber(): List<Int> {
         return numCreator.getNums()
