+++ src/test/kotlin/lotto/domain/LotteryShopTest.kt
@@ -3,6 +3,8 @@ package lotto.domain
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
+import org.junit.jupiter.params.ParameterizedTest
+import org.junit.jupiter.params.provider.CsvSource
 
 internal class LotteryShopTest {
     @Test
@@ -12,7 +14,7 @@ internal class LotteryShopTest {
 
     @Test
     internal fun `2장을 발급하면 로또 두장이 발급된다`() {
-        LotteryShop.getTickets(2).tickets.size shouldBe 2
+        LotteryShop.getTickets(2, listOf()).tickets.size shouldBe 2
     }
 
     @Test
@@ -20,5 +22,15 @@ internal class LotteryShopTest {
         assertThrows<IllegalArgumentException> {
             LotteryShop.validateManualNum(3, 2)
         }
+    }
+
+    @ParameterizedTest
+    @CsvSource(value = ["1,2,3,11,8,9;7;0.5", "1,2,3,4,8,9;7;5"], delimiter = ';')
+    internal fun `동일한 개수에 등수가 결정되고 그에 따른 수익률이 계산된다`(winningNums: String, bonus: Int, result: Float) {
+        val winningTicket = WinningTicket(Lotto(winningNums.split(",").map { it.toInt() }), bonus)
+        val testLotto = Tickets(1, listOf(), TestNumGenerator())
+        val score = winningTicket.score(testLotto)
+
+        LotteryShop.calculateRateOfReturn(10000, score) shouldBe result
     }
 }
